backend/scripts/audio_processor.py

import os
import numpy as np
import torch
import librosa
import soundfile as sf

class AudioProcessor:
    def __init__(self, model_dir):
        self.model_dir = model_dir
        self.denoise_model = None
        
    def load_denoise_model(self):
        """
        Load the audio denoising model
        """
        # TODO: Implement actual model loading
        # This is a placeholder for when you add your trained model
        model_path = os.path.join(self.model_dir, "audio_denoise", "model.pth")
        if os.path.exists(model_path):
            # self.denoise_model = torch.load(model_path)
            print(f"Audio denoising model loaded from {model_path}")
        else:
            print(f"Warning: No audio denoising model found at {model_path}")
    
    def denoise_audio(self, audio_path, output_path=None):
        """
        Denoise an audio file using the loaded model
        """
        # If no model is loaded, load it
        if self.denoise_model is None:
            self.load_denoise_model()
        
        # TODO: Replace with actual model inference
        # This is a placeholder implementation
        try:
            # Load audio file
            audio, sr = librosa.load(audio_path, sr=None)
            
            # Placeholder denoising operation (simple low-pass filter as an example)
            # This is NOT real denoising, just a demo
            from scipy.signal import butter, filtfilt
            
            def butter_lowpass(cutoff, fs, order=5):
                nyq = 0.5 * fs
                normal_cutoff = cutoff / nyq
                b, a = butter(order, normal_cutoff, btype='low', analog=False)
                return b, a

            def butter_lowpass_filter(data, cutoff, fs, order=5):
                b, a = butter_lowpass(cutoff, fs, order=order)
                y = filtfilt(b, a, data)
                return y
            
            # Filter design
            cutoff = 1000  # Cutoff frequency in Hz
            denoised_audio = butter_lowpass_filter(audio, cutoff, sr)
            
            # Save the result if output path is provided
            if output_path:
                sf.write(output_path, denoised_audio, sr)
                return output_path
            else:
                # Generate a default output path
                dir_name = os.path.dirname(audio_path)
                base_name = os.path.basename(audio_path)
                name, ext = os.path.splitext(base_name)
                output_path = os.path.join(dir_name, f"{name}_denoised{ext}")
                sf.write(output_path, denoised_audio, sr)
                return output_path
        except Exception as e:
            print(f"Error denoising audio: {e}")
            return None
