backend/scripts/image_processor.py

import os
import numpy as np
import torch
import cv2
from PIL import Image
from torchvision import transforms

class ImageProcessor:
    def __init__(self, model_dir):
        self.model_dir = model_dir
        self.denoise_model = None
        self.upscale_model = None
        
    def load_denoise_model(self):
        """
        Load the denoising model
        """
        # TODO: Implement actual model loading
        # This is a placeholder for when you add your trained model
        model_path = os.path.join(self.model_dir, "image_denoise", "model.pth")
        if os.path.exists(model_path):
            # self.denoise_model = torch.load(model_path)
            print(f"Denoising model loaded from {model_path}")
        else:
            print(f"Warning: No denoising model found at {model_path}")
    
    def load_upscale_model(self):
        """
        Load the upscaling model
        """
        # TODO: Implement actual model loading
        # This is a placeholder for when you add your trained model
        model_path = os.path.join(self.model_dir, "image_upscale", "model.pth")
        if os.path.exists(model_path):
            # self.upscale_model = torch.load(model_path)
            print(f"Upscaling model loaded from {model_path}")
        else:
            print(f"Warning: No upscaling model found at {model_path}")
    
    def denoise_image(self, image_path, output_path=None):
        """
        Denoise an image using the loaded model
        """
        # If no model is loaded, load it
        if self.denoise_model is None:
            self.load_denoise_model()
        
        # TODO: Replace with actual model inference
        # This is a placeholder implementation
        try:
            # Load and process the image
            img = cv2.imread(image_path)
            if img is None:
                raise ValueError(f"Failed to load image at {image_path}")
                
            # Placeholder denoising operation (Gaussian blur as a simple example)
            denoised = cv2.GaussianBlur(img, (5, 5), 0)
            
            # Save the result if output path is provided
            if output_path:
                cv2.imwrite(output_path, denoised)
                return output_path
            else:
                # Generate a default output path
                dir_name = os.path.dirname(image_path)
                base_name = os.path.basename(image_path)
                name, ext = os.path.splitext(base_name)
                output_path = os.path.join(dir_name, f"{name}_denoised{ext}")
                cv2.imwrite(output_path, denoised)
                return output_path
        except Exception as e:
            print(f"Error denoising image: {e}")
            return None
    
    def upscale_image(self, image_path, output_path=None, scale_factor=2):
        """
        Upscale an image using the loaded model
        """
        # If no model is loaded, load it
        if self.upscale_model is None:
            self.load_upscale_model()
        
        # TODO: Replace with actual model inference
        # This is a placeholder implementation
        try:
            # Load and process the image
            img = cv2.imread(image_path)
            if img is None:
                raise ValueError(f"Failed to load image at {image_path}")
                
            # Placeholder upscaling operation (simple resize as an example)
            height, width = img.shape[:2]
            upscaled = cv2.resize(img, (width * scale_factor, height * scale_factor), 
                                interpolation=cv2.INTER_CUBIC)
            
            # Save the result if output path is provided
            if output_path:
                cv2.imwrite(output_path, upscaled)
                return output_path
            else:
                # Generate a default output path
                dir_name = os.path.dirname(image_path)
                base_name = os.path.basename(image_path)
                name, ext = os.path.splitext(base_name)
                output_path = os.path.join(dir_name, f"{name}_upscaled{ext}")
                cv2.imwrite(output_path, upscaled)
                return output_path
        except Exception as e:
            print(f"Error upscaling image: {e}")
            return None
