backend/requirements.txt

# Core packages
numpy==1.24.3
scipy==1.10.1
Pillow==9.5.0
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2

# API and server
fastapi==0.95.2
uvicorn==0.22.0
python-multipart==0.0.6

# Image processing
opencv-python==4.7.0.72
scikit-image==0.20.0

# Audio processing
librosa==0.10.0.post2

# Utilities
tqdm==4.65.0
